Clean and rename utils
